(sentence building) More progress on the data handling side of the dialogue generator
Still WIP.
Added a way to prompt for data (ex: ask a name, ask gender, ask for a random piece of knowledge, etc).
Character can now be prompted for name, gender and other pieces of data. Most are unhandled so far. The plan is to handle each type of data in a slightly different way.
Next up is:
-Editing sentence building to have less straightforward notions: what you are, what you have, what you own and what you do should be expressed differently.
-Editing sentence building to use its own data format instead of the Fact struct it was built around. This is to be able to give more nuances to sentences, notably by using the above new notions.
-And finally handle all types of data and convert them to that format to be passed to the sentence builder.
Once that is all done, everything will be expressable, and thought will have to be given to the overall system, including what should be expressed when and maintaining a coherent flow of monologue.